Fix migration execution - Version20201008101817 (#13526)
* Fix migration execution - Version20201008101817 follow up to #13438

* Fix migration execution - Version20201008101817, review changes - follow up to #13438

+++ bundles/CoreBundle/Migrations/Version20201008101817.php
@@ -29,10 +29,8 @@ final class Version20201008101817 extends AbstractMigration
      */
     public function up(Schema $schema): void
     {
-        try {
+        if (array_key_exists('documents_editables', $this->connection->getSchemaManager()->listViews())) {
             $this->addSql('DROP VIEW IF EXISTS documents_editables;');
-        } catch (Exception $e) {
-            $this->write('Could not drop view: ' . $e->getMessage());
         }
 
         if ($schema->hasTable('documents_elements')) {
